refactor to prepare for timeLine import
